BozorShopPrototype-main/requirements.txt
django==6.0.3
djangorestframework==3.17.1
django-filter==25.2
